+++ fitsio/fitsio_pywrap.c
@@ -34,9 +34,16 @@
 // not sure where this is defined in numpy...
 #define NUMPY_MAX_DIMS 32
 
+// max len of python error message
+#define PYFITS_ERRMSG_LEN 1024
+
 struct PyFITSObject {
     PyObject_HEAD
     fitsfile* fits;
+    // we store the python error message here so that we record all error
+    // messages as they happen. sometimes cfitsio will clear
+    // the error stack and this removes important debugging info
+    char pyfits_errmsg[PYFITS_ERRMSG_LEN];
 };
 
 #ifdef FITSIO_PYWRAP_ALWAYS_NONSTANDARD_STRINGS
@@ -228,26 +235,60 @@ static char* get_object_as_string(PyObject* obj)
     return strdata;
 }
 
-static void set_ioerr_string_from_status(int status) {
+static void set_ioerr_string_from_status(int status, struct PyFITSObject* self) {
     char status_str[FLEN_STATUS], errmsg[FLEN_ERRMSG];
-    char message[1024];
+    char str_with_newline[1024];
+    int nleft;
+    // sometimes this function is called without an instantiated PyFITSPObject.
+    // in that case, we use a local buffer and grab all of the information we can
+    // to set the error string. this edge case means we will sometimes miss error
+    // messages that cfitsio puts on the stack but then removes later.
+    char message_if_null[PYFITS_ERRMSG_LEN];
 
-    int nleft=1024;
+    // pointer to the error message buffer we are using
+    char* message;
+
+    // always init error message to zero length string
+    // init of self->pyfits_errmsg is done when object is created
+    message_if_null[0] = '\0';
+
+    if (self != NULL) {
+        message = &(self->pyfits_errmsg[0]);
+    } else {
+        message = &(message_if_null[0]);
+    }
+
+    // the number of characters left is the number of characters not
+    // used in the message variable minus one. We remove one for the
+    // C null termination character. So if the variable is PYFITS_ERRMSG_LEN
+    // characters long, then nleft is PYFITS_ERRMSG_LEN - 1.
+    // This works because:
+    //   - strncat always adds the null termination character
+    //   - strlen returns the number of characters up to the first null (and
+    //     does not include the null itself in the total count).
+    // Thus we never have to account for the null ourselves except in the initial
+    // number of characters we have to use for holding error messages (which we
+    // set to one less than the total storage we have so strncat always has
+    // room for the null.)
+    nleft = PYFITS_ERRMSG_LEN - strlen(message) - 1;
 
     if (status) {
         fits_get_errstatus(status, status_str);  /* get the error description */
 
-        sprintf(message, "FITSIO status = %d: %s\n", status, status_str);
-
-        nleft -= strlen(status_str)+1;
+        sprintf(str_with_newline, "FITSIO status = %d: %s\n", status, status_str);
+        if (nleft >= strlen(str_with_newline)) {
+            strncat(message, str_with_newline, nleft);
+            nleft -= strlen(str_with_newline);
+        }
 
         while ( nleft > 0 && fits_read_errmsg(errmsg) )  { /* get error stack messages */
-            strncat(message, errmsg, nleft-1);
-            nleft -= strlen(errmsg)+1;
-            if (nleft >= 2) {
-                strncat(message, "\n", nleft-1);
+            sprintf(str_with_newline, "%s\n", errmsg);
+            if (nleft >= strlen(str_with_newline)) {
+                strncat(message, str_with_newline, nleft);
+                nleft -= strlen(str_with_newline);
+            } else {
+                break;
             }
-            nleft-=2;
         }
         PyErr_SetString(PyExc_IOError, message);
     }
@@ -440,6 +481,9 @@ PyFITSObject_init(struct PyFITSObject* self, PyObject *args, PyObject *kwds)
     int status=0;
     int create=0;
 
+    // init the error message to an empty string
+    self->pyfits_errmsg[0] = '\0';
+
     if (!PyArg_ParseTuple(args, (char*)"sii", &filename, &mode, &create)) {
         return -1;
     }
@@ -447,12 +491,12 @@ PyFITSObject_init(struct PyFITSObject* self, PyObject *args, PyObject *kwds)
     if (create) {
         // create and open
         if (fits_create_file(&self->fits, filename, &status)) {
-            set_ioerr_string_from_status(status);
+            set_ioerr_string_from_status(status, self);
             return -1;
         }
     } else {
         if (fits_open_file(&self->fits, filename, mode, &status)) {
-            set_ioerr_string_from_status(status);
+            set_ioerr_string_from_status(status, self);
             return -1;
         }
     }
@@ -470,7 +514,7 @@ PyFITSObject_repr(struct PyFITSObject* self) {
         char repr[2056];
 
         if (fits_file_name(self->fits, filename, &status)) {
-            set_ioerr_string_from_status(status);
+            set_ioerr_string_from_status(status, self);
             return NULL;
         }
 
@@ -489,7 +533,7 @@ PyFITSObject_filename(struct PyFITSObject* self) {
         char filename[FLEN_FILENAME];
         PyObject* fnameObj=NULL;
         if (fits_file_name(self->fits, filename, &status)) {
-            set_ioerr_string_from_status(status);
+            set_ioerr_string_from_status(status, self);
             return NULL;
         }
 
@@ -510,7 +554,7 @@ PyFITSObject_close(struct PyFITSObject* self)
     if (fits_close_file(self->fits, &status)) {
         self->fits=NULL;
         /*
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
         */
     }
@@ -601,7 +645,7 @@ PyFITSObject_movnam_hdu(struct PyFITSObject* self, PyObject* args) {
     }
 
     if (fits_movnam_hdu(self->fits, hdutype, extname,  extver, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -627,7 +671,7 @@ PyFITSObject_movabs_hdu(struct PyFITSObject* self, PyObject* args) {
     }
 
     if (fits_movabs_hdu(self->fits, hdunum, &hdutype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
     hdutypeObj = PyLong_FromLong((long)hdutype);
@@ -659,7 +703,7 @@ PyFITSObject_get_hdu_info(struct PyFITSObject* self, PyObject* args) {
     }
 
     if (fits_movabs_hdu(self->fits, hdunum, &hdutype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -954,7 +998,7 @@ PyFITSObject_get_hdu_name_version(struct PyFITSObject* self, PyObject* args) {
     }
 
     if (fits_movabs_hdu(self->fits, hdunum, &hdutype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -1295,7 +1339,7 @@ static int create_empty_hdu(struct PyFITSObject* self)
     int naxis=0;
     long* naxes=NULL;
     if (fits_create_img(self->fits, bitpix, naxis, naxes, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return 1;
     }
 
@@ -1312,7 +1356,8 @@ static int create_empty_hdu(struct PyFITSObject* self)
 //
 // note tile dims are written reverse order since
 // python orders C and fits orders Fortran
-static int set_compression(fitsfile *fits,
+static int set_compression(struct PyFITSObject* self,
+                           fitsfile *fits,
                            int comptype,
                            PyObject* tile_dims_obj,
                            int *status) {
@@ -1323,7 +1368,7 @@ static int set_compression(fitsfile *fits,
 
     // can be NOCOMPRESS (0)
     if (fits_set_compression_type(fits, comptype, status)) {
-        set_ioerr_string_from_status(*status);
+        set_ioerr_string_from_status(*status, self);
         goto _set_compression_bail;
         return 1;
     }
@@ -1415,7 +1460,7 @@ PyFITSObject_create_image_hdu(struct PyFITSObject* self, PyObject* args, PyObjec
     // We do allow overriding the dither seed and then putting it back
     // Luckily cfitsio has public APIs for that so this should be robust
     if (fits_get_dither_seed(self->fits, &orig_dither_seed, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         goto create_image_hdu_cleanup;
     } else {
         // tell the code to put back the dither seed later
@@ -1532,28 +1577,28 @@ PyFITSObject_create_image_hdu(struct PyFITSObject* self, PyObject* args, PyObjec
 
         if ((comptype_obj != Py_None) || (tile_dims_obj != Py_None)) {
             // exception strings are set internally
-            if (set_compression(self->fits, comptype, tile_dims_obj, &status)) {
+            if (set_compression(self, self->fits, comptype, tile_dims_obj, &status)) {
                 goto create_image_hdu_cleanup;
             }
         }
 
         if (qlevel_obj != Py_None) {
             if (fits_set_quantize_level(self->fits, qlevel, &status)) {
-                set_ioerr_string_from_status(status);
+                set_ioerr_string_from_status(status, self);
                 goto create_image_hdu_cleanup;
             }
         }
 
         if (qmethod_obj != Py_None) {
             if (fits_set_quantize_method(self->fits, qmethod, &status)) {
-                set_ioerr_string_from_status(status);
+                set_ioerr_string_from_status(status, self);
                 goto create_image_hdu_cleanup;
             }
         }
 
         if (dither_seed_obj != Py_None) {
             if (fits_set_dither_seed(self->fits, dither_seed, &status)) {
-                set_ioerr_string_from_status(status);
+                set_ioerr_string_from_status(status, self);
                 goto create_image_hdu_cleanup;
             }
         }
@@ -1561,21 +1606,21 @@ PyFITSObject_create_image_hdu(struct PyFITSObject* self, PyObject* args, PyObjec
         if (comptype == HCOMPRESS_1) {
             if (hcomp_scale_obj != Py_None) {
                 if (fits_set_hcomp_scale(self->fits, hcomp_scale, &status)) {
-                    set_ioerr_string_from_status(status);
+                    set_ioerr_string_from_status(status, self);
                     goto create_image_hdu_cleanup;
                 }
             }
 
             if (hcomp_smooth_obj != Py_None) {
                 if (fits_set_hcomp_smooth(self->fits, hcomp_smooth, &status)) {
-                    set_ioerr_string_from_status(status);
+                    set_ioerr_string_from_status(status, self);
                     goto create_image_hdu_cleanup;
                 }
             }
         }
 
         if (fits_create_img(self->fits, image_datatype, ndims, dims, &status)) {
-            set_ioerr_string_from_status(status);
+            set_ioerr_string_from_status(status, self);
             goto create_image_hdu_cleanup;
         }
     }
@@ -1584,12 +1629,12 @@ PyFITSObject_create_image_hdu(struct PyFITSObject* self, PyObject* args, PyObjec
 
             // comments are NULL
             if (fits_update_key_str(self->fits, "EXTNAME", extname, NULL, &status)) {
-                set_ioerr_string_from_status(status);
+                set_ioerr_string_from_status(status, self);
                 goto create_image_hdu_cleanup;
             }
             if (extver > 0) {
                 if (fits_update_key_lng(self->fits, "EXTVER", (LONGLONG) extver, NULL, &status)) {
-                    set_ioerr_string_from_status(status);
+                    set_ioerr_string_from_status(status, self);
                     goto create_image_hdu_cleanup;
                 }
             }
@@ -1598,7 +1643,7 @@ PyFITSObject_create_image_hdu(struct PyFITSObject* self, PyObject* args, PyObjec
 
     if (nkeys > 0) {
         if (fits_set_hdrsize(self->fits, nkeys, &status) ) {
-            set_ioerr_string_from_status(status);
+            set_ioerr_string_from_status(status, self);
             goto create_image_hdu_cleanup;
         }
     }
@@ -1610,14 +1655,14 @@ PyFITSObject_create_image_hdu(struct PyFITSObject* self, PyObject* args, PyObjec
         nelements = PyArray_SIZE(array);
         data = PyArray_DATA(array);
         if (fits_write_img(self->fits, datatype, firstpixel, nelements, data, &status)) {
-            set_ioerr_string_from_status(status);
+            set_ioerr_string_from_status(status, self);
             goto create_image_hdu_cleanup;
         }
     }
 
     // this flushes all buffers
     if (fits_flush_file(self->fits, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         goto create_image_hdu_cleanup;
     }
 
@@ -1632,14 +1677,14 @@ create_image_hdu_cleanup:
         || (hcomp_smooth_obj != Py_None)
     ) {
         if (fits_unset_compression_request(self->fits, &status)) {
-            set_ioerr_string_from_status(status);
+            set_ioerr_string_from_status(status, self);
         }
     }
 
     // put back the dither seed after we reset any compression params
     if (got_dither_seed == 1) {
         if (fits_set_dither_seed(self->fits, orig_dither_seed, &status)) {
-            set_ioerr_string_from_status(status);
+            set_ioerr_string_from_status(status, self);
         }
     }
 
@@ -1678,13 +1723,13 @@ PyFITSObject_reshape_image(struct PyFITSObject* self, PyObject* args) {
     dims_array = (PyArrayObject *) dims_obj;
 
     if (fits_movabs_hdu(self->fits, hdunum, &hdutype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
     // existing image params, just to get bitpix
     if (fits_get_img_paramll(self->fits, maxdim, &bitpix, &ndims_orig, dims_orig, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -1695,7 +1740,7 @@ PyFITSObject_reshape_image(struct PyFITSObject* self, PyObject* args) {
     }
 
     if (fits_resize_imgll(self->fits, bitpix, ndims, dims, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -1731,7 +1776,7 @@ PyFITSObject_write_image(struct PyFITSObject* self, PyObject* args) {
     array = (PyArrayObject *) array_obj;
 
     if (fits_movabs_hdu(self->fits, hdunum, &hdutype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -1750,13 +1795,104 @@ PyFITSObject_write_image(struct PyFITSObject* self, PyObject* args) {
     nelements = PyArray_SIZE(array);
     firstpixel = (LONGLONG) firstpixel_py;
     if (fits_write_img(self->fits, datatype, firstpixel, nelements, data, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
     // this is a full file close and reopen
     if (fits_flush_file(self->fits, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
+        return NULL;
+    }
+
+    Py_RETURN_NONE;
+}
+
+
+// write a rectangular subset to the image in an existing HDU
+// created using create_image_hdu
+// dims are not checked
+static PyObject *
+PyFITSObject_write_subset(struct PyFITSObject* self, PyObject* args) {
+    int hdunum=0;
+    int hdutype=0;
+    int image_datatype=0; // fits type for image, AKA bitpix
+    int datatype=0; // type for the data we entered
+
+    long fpixel[CFITSIO_MAX_ARRAY_DIMS]={0};
+    long lpixel[CFITSIO_MAX_ARRAY_DIMS]={0};
+    npy_int64 pixval=0;
+    npy_intp i=0;
+
+    PyObject* array_obj=NULL;
+    PyObject* firstpixel_obj=NULL;
+    PyObject* lastpixel_obj=NULL;
+    PyArrayObject* array=NULL;
+    PyArrayObject* firstpixel=NULL;
+    PyArrayObject* lastpixel=NULL;
+    void* data=NULL;
+    int npy_dtype=0, ndims;
+    int status=0;
+
+    if (self->fits == NULL) {
+        PyErr_SetString(PyExc_ValueError, "fits file is NULL");
+        return NULL;
+    }
+
+    if (!PyArg_ParseTuple(args, (char*)"iOOO", &hdunum, &array_obj, &firstpixel_obj, &lastpixel_obj)) {
+        return NULL;
+    }
+
+    if (fits_movabs_hdu(self->fits, hdunum, &hdutype, &status)) {
+        set_ioerr_string_from_status(status, self);
+        return NULL;
+    }
+
+    array = (PyArrayObject *) array_obj;
+    if (!PyArray_Check(array)) {
+        PyErr_SetString(PyExc_TypeError, "input data must be an array.");
+        return NULL;
+    }
+    npy_dtype = PyArray_TYPE(array);
+    if (npy_to_fits_image_types(npy_dtype, &image_datatype, &datatype)) {
+        return NULL;
+    }
+
+    firstpixel = (PyArrayObject *) firstpixel_obj;
+    if (!PyArray_Check(firstpixel)) {
+        PyErr_SetString(PyExc_TypeError, "input firstpixel must be an array.");
+        return NULL;
+    }
+
+    lastpixel = (PyArrayObject *) lastpixel_obj;
+    if (!PyArray_Check(lastpixel)) {
+        PyErr_SetString(PyExc_TypeError, "input lastpixel must be an array.");
+        return NULL;
+    }
+
+    if (PyArray_SIZE(firstpixel) != PyArray_SIZE(lastpixel)) {
+        PyErr_SetString(PyExc_TypeError, "input firstpixel and lastpixel must have the same size.");
+        return NULL;
+    }
+
+    ndims = PyArray_SIZE(firstpixel);
+    for (i=0; i<ndims; i++) {
+        pixval = *(npy_int64 *) PyArray_GETPTR1(firstpixel, i);
+        fpixel[i] = (long) pixval;
+
+        pixval = *(npy_int64 *) PyArray_GETPTR1(lastpixel, i);
+        lpixel[i] = (long) pixval;
+    }
+
+    data = PyArray_DATA(array);
+    if (fits_write_subset(self->fits, datatype, fpixel, lpixel, data, &status)) {
+        set_ioerr_string_from_status(status, self);
+        return NULL;
+    }
+
+    // this is a full file close and reopen
+    if (fits_flush_file(self->fits, &status)) {
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -1771,7 +1907,7 @@ PyFITSObject_write_image(struct PyFITSObject* self, PyObject* args) {
  * The keys are written as TDIM{colnum}
  */
 static int
-add_tdims_from_listobj(fitsfile* fits, PyObject* tdimObj, int ncols) {
+add_tdims_from_listobj(struct PyFITSObject* self, fitsfile* fits, PyObject* tdimObj, int ncols) {
     int status=0, i=0;
     size_t size=0;
     char keyname[20];
@@ -1811,7 +1947,7 @@ add_tdims_from_listobj(fitsfile* fits, PyObject* tdimObj, int ncols) {
             free(tdim);
 
             if (status) {
-                set_ioerr_string_from_status(status);
+                set_ioerr_string_from_status(status, self);
                 return 1;
             }
         }
@@ -1881,11 +2017,11 @@ PyFITSObject_create_table_hdu(struct PyFITSObject* self, PyObject* args, PyObjec
     nfields = ttyp->size;
     if ( fits_create_tbl(self->fits, table_type, nrows, nfields,
                          ttyp->data, tform->data, tunit->data, extname_use, &status) ) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         goto create_table_cleanup;
     }
 
-    if (add_tdims_from_listobj(self->fits, tdimObj, nfields)) {
+    if (add_tdims_from_listobj(self, self->fits, tdimObj, nfields)) {
         status=99;
         goto create_table_cleanup;
     }
@@ -1894,7 +2030,7 @@ PyFITSObject_create_table_hdu(struct PyFITSObject* self, PyObject* args, PyObjec
         if (extver > 0) {
 
             if (fits_update_key_lng(self->fits, "EXTVER", (LONGLONG) extver, NULL, &status)) {
-                set_ioerr_string_from_status(status);
+                set_ioerr_string_from_status(status, self);
                 goto create_table_cleanup;
             }
         }
@@ -1902,14 +2038,14 @@ PyFITSObject_create_table_hdu(struct PyFITSObject* self, PyObject* args, PyObjec
 
     if (nkeys > 0) {
         if (fits_set_hdrsize(self->fits, nkeys, &status) ) {
-            set_ioerr_string_from_status(status);
+            set_ioerr_string_from_status(status, self);
             goto create_table_cleanup;
         }
     }
 
     // this does a full close and reopen
     if (fits_flush_file(self->fits, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         goto create_table_cleanup;
     }
 
@@ -1950,12 +2086,12 @@ PyFITSObject_insert_col(struct PyFITSObject* self, PyObject* args, PyObject* kwd
     }
 
     if (fits_movabs_hdu(self->fits, hdunum, &hdutype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
     if (fits_insert_col(self->fits, colnum, ttype, tform, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -1973,14 +2109,14 @@ PyFITSObject_insert_col(struct PyFITSObject* self, PyObject* args, PyObject* kwd
         free(tdim);
 
         if (status) {
-            set_ioerr_string_from_status(status);
+            set_ioerr_string_from_status(status, self);
             return NULL;
         }
     }
 
     // this does a full close and reopen
     if (fits_flush_file(self->fits, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -1993,6 +2129,7 @@ PyFITSObject_insert_col(struct PyFITSObject* self, PyObject* args, PyObject* kwd
 // No error checking performed here
 static
 int write_string_column(
+        struct PyFITSObject* self,
         fitsfile *fits,  /* I - FITS file pointer                       */
         int  colnum,     /* I - number of column to write (1 = 1st col) */
         LONGLONG  firstrow,  /* I - first row to write (1 = 1st row)        */
@@ -2024,7 +2161,7 @@ int write_string_column(
     }
 
     if( fits_write_col_str(fits, colnum, firstrow, firstelem, nelem, strdata, status)) {
-        set_ioerr_string_from_status(*status);
+        set_ioerr_string_from_status(*status, self);
         free(strdata);
         return 1;
     }
@@ -2071,7 +2208,7 @@ PyFITSObject_write_column(struct PyFITSObject* self, PyObject* args, PyObject* k
     firstrow = (LONGLONG) firstrow_py;
 
     if (fits_movabs_hdu(self->fits, hdunum, &hdutype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -2109,25 +2246,25 @@ PyFITSObject_write_column(struct PyFITSObject* self, PyObject* args, PyObject* k
     if (fits_dtype == TSTRING) {
 
         // this is my wrapper for strings
-        if (write_string_column(self->fits, colnum, firstrow, firstelem, nelem, data, &status)) {
-            set_ioerr_string_from_status(status);
+        if (write_string_column(self, self->fits, colnum, firstrow, firstelem, nelem, data, &status)) {
+            set_ioerr_string_from_status(status, self);
             return NULL;
         }
     } else if (fits_dtype == TBIT) {
         if (fits_write_col_bit(self->fits, colnum, firstrow, firstelem, nelem, data, &status)) {
-            set_ioerr_string_from_status(status);
+            set_ioerr_string_from_status(status, self);
             return NULL;
         }
     } else {
         if( fits_write_col(self->fits, fits_dtype, colnum, firstrow, firstelem, nelem, data, &status)) {
-            set_ioerr_string_from_status(status);
+            set_ioerr_string_from_status(status, self);
             return NULL;
         }
     }
 
     // this is a full file close and reopen
     if (fits_flush_file(self->fits, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -2178,7 +2315,7 @@ PyFITSObject_write_columns(struct PyFITSObject* self, PyObject* args, PyObject* 
     firstrow = (LONGLONG) firstrow_py;
 
     if (fits_movabs_hdu(self->fits, hdunum, &hdutype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -2269,14 +2406,15 @@ PyFITSObject_write_columns(struct PyFITSObject* self, PyObject* args, PyObject* 
         for (icol=0; icol<ncols; icol++) {
             data=PyArray_GETPTR1(array_ptrs[icol], irow);
             if (is_string[icol]) {
-                if (write_string_column(self->fits,
+                if (write_string_column(self,
+                                        self->fits,
                                         colnums[icol],
                                         thisrow,
                                         firstelem,
                                         nperrow[icol],
                                         (char*)data,
                                         &status)) {
-                    set_ioerr_string_from_status(status);
+                    set_ioerr_string_from_status(status, self);
                     goto _fitsio_pywrap_write_columns_bail;
                 }
 
@@ -2288,7 +2426,7 @@ PyFITSObject_write_columns(struct PyFITSObject* self, PyObject* args, PyObject* 
                                        nperrow[icol],
                                        data,
                                        &status)) {
-                    set_ioerr_string_from_status(status);
+                    set_ioerr_string_from_status(status, self);
                     goto _fitsio_pywrap_write_columns_bail;
                 }
             } else {
@@ -2300,7 +2438,7 @@ PyFITSObject_write_columns(struct PyFITSObject* self, PyObject* args, PyObject* 
                                    nperrow[icol],
                                    data,
                                    &status)) {
-                    set_ioerr_string_from_status(status);
+                    set_ioerr_string_from_status(status, self);
                     goto _fitsio_pywrap_write_columns_bail;
                 }
             }
@@ -2312,21 +2450,21 @@ PyFITSObject_write_columns(struct PyFITSObject* self, PyObject* args, PyObject* 
     if (fits_dtype == TSTRING) {
 
         // this is my wrapper for strings
-        if (write_string_column(self->fits, colnum, firstrow, firstelem, nelem, data, &status)) {
-            set_ioerr_string_from_status(status);
+        if (write_string_column(self, self->fits, colnum, firstrow, firstelem, nelem, data, &status)) {
+            set_ioerr_string_from_status(status, self);
             return NULL;
         }
 
     } else {
         if( fits_write_col(self->fits, fits_dtype, colnum, firstrow, firstelem, nelem, data, &status)) {
-            set_ioerr_string_from_status(status);
+            set_ioerr_string_from_status(status, self);
             return NULL;
         }
     }
 
     // this is a full file close and reopen
     if (fits_flush_file(self->fits, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
     */
@@ -2404,6 +2542,7 @@ write_var_string_column_cleanup:
  */
 static
 int write_var_num_column(
+        struct PyFITSObject* self,
         fitsfile *fits,  /* I - FITS file pointer                       */
         int  colnum,     /* I - number of column to write (1 = 1st col) */
         LONGLONG  firstrow,  /* I - first row to write (1 = 1st row)        */
@@ -2457,7 +2596,7 @@ int write_var_num_column(
         Py_XDECREF(el_array);
 
         if(res > 0) {
-            set_ioerr_string_from_status(*status);
+            set_ioerr_string_from_status(*status, self);
             return 1;
         }
     }
@@ -2505,7 +2644,7 @@ PyFITSObject_write_var_column(struct PyFITSObject* self, PyObject* args, PyObjec
     firstrow = (LONGLONG) firstrow_py;
 
     if (fits_movabs_hdu(self->fits, hdunum, &hdutype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -2525,27 +2664,27 @@ PyFITSObject_write_var_column(struct PyFITSObject* self, PyObject* args, PyObjec
     // determine the fits dtype for this column.  We will use this to get data
     // from the array for writing
     if (fits_get_eqcoltypell(self->fits, colnum, &fits_dtype, NULL, NULL, &status) > 0) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
     if (fits_dtype == -TSTRING) {
         if (write_var_string_column(self->fits, colnum, firstrow, array, &status)) {
             if (status != 0) {
-                set_ioerr_string_from_status(status);
+                set_ioerr_string_from_status(status, self);
             }
             return NULL;
         }
     } else {
-        if (write_var_num_column(self->fits, colnum, firstrow, fits_dtype, array, &status)) {
-            set_ioerr_string_from_status(status);
+        if (write_var_num_column(self, self->fits, colnum, firstrow, fits_dtype, array, &status)) {
+            set_ioerr_string_from_status(status, self);
             return NULL;
         }
     }
 
     // this is a full file close and reopen
     if (fits_flush_file(self->fits, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -2575,7 +2714,7 @@ PyFITSObject_write_record(struct PyFITSObject* self, PyObject* args) {
         return NULL;
     }
     if (fits_movabs_hdu(self->fits, hdunum, &hdutype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
     // card not null terminated, so we copy everything.  GCC will
@@ -2583,13 +2722,13 @@ PyFITSObject_write_record(struct PyFITSObject* self, PyObject* args) {
     strncpy(card, cardin, FLEN_CARD);
 
     if (fits_write_record(self->fits, card, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
     // this does not close and reopen
     if (fits_flush_buffer(self->fits, 0, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -2618,7 +2757,7 @@ PyFITSObject_write_string_key(struct PyFITSObject* self, PyObject* args) {
         return NULL;
     }
     if (fits_movabs_hdu(self->fits, hdunum, &hdutype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -2627,13 +2766,13 @@ PyFITSObject_write_string_key(struct PyFITSObject* self, PyObject* args) {
     }
 
     if (fits_update_key_longstr(self->fits, keyname, value, comment, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
     // this does not close and reopen
     if (fits_flush_buffer(self->fits, 0, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -2662,7 +2801,7 @@ PyFITSObject_write_double_key(struct PyFITSObject* self, PyObject* args) {
         return NULL;
     }
     if (fits_movabs_hdu(self->fits, hdunum, &hdutype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -2671,13 +2810,13 @@ PyFITSObject_write_double_key(struct PyFITSObject* self, PyObject* args) {
     }
 
     if (fits_update_key_dbl(self->fits, keyname, value, decimals, comment, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
     // this does not close and reopen
     if (fits_flush_buffer(self->fits, 0, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -2705,7 +2844,7 @@ PyFITSObject_write_long_long_key(struct PyFITSObject* self, PyObject* args) {
         return NULL;
     }
     if (fits_movabs_hdu(self->fits, hdunum, &hdutype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -2714,13 +2853,13 @@ PyFITSObject_write_long_long_key(struct PyFITSObject* self, PyObject* args) {
     }
 
     if (fits_update_key_lng(self->fits, keyname, (LONGLONG) value, comment, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
     // this does not close and reopen
     if (fits_flush_buffer(self->fits, 0, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -2747,7 +2886,7 @@ PyFITSObject_write_logical_key(struct PyFITSObject* self, PyObject* args) {
         return NULL;
     }
     if (fits_movabs_hdu(self->fits, hdunum, &hdutype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -2756,13 +2895,13 @@ PyFITSObject_write_logical_key(struct PyFITSObject* self, PyObject* args) {
     }
 
     if (fits_update_key_log(self->fits, keyname, value, comment, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
     // this does not close and reopen
     if (fits_flush_buffer(self->fits, 0, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -2787,18 +2926,18 @@ PyFITSObject_write_comment(struct PyFITSObject* self, PyObject* args) {
         return NULL;
     }
     if (fits_movabs_hdu(self->fits, hdunum, &hdutype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
     if (fits_write_comment(self->fits, comment, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
     // this does not close and reopen
     if (fits_flush_buffer(self->fits, 0, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -2823,18 +2962,18 @@ PyFITSObject_write_history(struct PyFITSObject* self, PyObject* args) {
         return NULL;
     }
     if (fits_movabs_hdu(self->fits, hdunum, &hdutype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
     if (fits_write_history(self->fits, history, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
     // this does not close and reopen
     if (fits_flush_buffer(self->fits, 0, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -2890,18 +3029,18 @@ PyFITSObject_write_continue(struct PyFITSObject* self, PyObject* args) {
         return NULL;
     }
     if (fits_movabs_hdu(self->fits, hdunum, &hdutype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
     if (fits_write_continue(self->fits, value, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
     // this does not close and reopen
     if (fits_flush_buffer(self->fits, 0, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -2928,7 +3067,7 @@ PyFITSObject_write_undefined_key(struct PyFITSObject* self, PyObject* args) {
         return NULL;
     }
     if (fits_movabs_hdu(self->fits, hdunum, &hdutype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -2937,13 +3076,13 @@ PyFITSObject_write_undefined_key(struct PyFITSObject* self, PyObject* args) {
     }
 
     if (fits_update_key_null(self->fits, keyname, comment, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
     // this does not close and reopen
     if (fits_flush_buffer(self->fits, 0, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -2984,18 +3123,18 @@ PyFITSObject_insert_rows(struct PyFITSObject* self, PyObject* args, PyObject* kw
     }
 
     if (fits_movabs_hdu(self->fits, hdunum, &hdutype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
     if (fits_insert_rows(self->fits, firstrow, nrows, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
     // this does a full close and reopen
     if (fits_flush_file(self->fits, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -3040,18 +3179,18 @@ PyFITSObject_delete_row_range(struct PyFITSObject* self, PyObject* args, PyObjec
     }
 
     if (fits_movabs_hdu(self->fits, hdunum, &hdutype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
     if (fits_delete_rows(self->fits, slice_start, nrows, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
     // this does a full close and reopen
     if (fits_flush_file(self->fits, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -3093,18 +3232,18 @@ PyFITSObject_delete_rows(struct PyFITSObject* self, PyObject* args, PyObject* kw
     }
 
     if (fits_movabs_hdu(self->fits, hdunum, &hdutype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
     if (fits_delete_rowlistll(self->fits, rows, nrows, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
     // this does a full close and reopen
     if (fits_flush_file(self->fits, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -3376,7 +3515,7 @@ PyFITSObject_read_column(struct PyFITSObject* self, PyObject* args) {
         return NULL;
     }
     if (fits_movabs_hdu(self->fits, hdunum, &hdutype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -3394,7 +3533,7 @@ PyFITSObject_read_column(struct PyFITSObject* self, PyObject* args) {
 
     if (hdutype == ASCII_TBL) {
         if (read_ascii_column(self->fits, colnum, array, rows_array, sortind_array, &status)) {
-            set_ioerr_string_from_status(status);
+            set_ioerr_string_from_status(status, self);
             return NULL;
         }
     } else {
@@ -3410,7 +3549,7 @@ PyFITSObject_read_column(struct PyFITSObject* self, PyObject* args) {
         }
 
         if (read_binary_column(self->fits, colnum, nrows, rows, sortind, data, stride, &status)) {
-            set_ioerr_string_from_status(status);
+            set_ioerr_string_from_status(status, self);
             return NULL;
         }
     }
@@ -3540,7 +3679,7 @@ PyFITSObject_read_var_column_as_list(struct PyFITSObject* self, PyObject* args) 
         return NULL;
     }
     if (fits_movabs_hdu(self->fits, hdunum, &hdutype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -3556,7 +3695,7 @@ PyFITSObject_read_var_column_as_list(struct PyFITSObject* self, PyObject* args) 
     }
 
     if (fits_get_coltypell(self->fits, colnum, &fits_dtype, &repeat, &width, &status) > 0) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -3618,7 +3757,7 @@ read_var_column_cleanup:
         Py_XDECREF(tempObj);
         free_all_python_list(listObj);
         if (status != 0) {
-            set_ioerr_string_from_status(status);
+            set_ioerr_string_from_status(status, self);
         }
         return NULL;
     }
@@ -3760,7 +3899,7 @@ PyFITSObject_read_columns_as_rec(struct PyFITSObject* self, PyObject* args) {
 recread_columns_cleanup:
 
     if (status != 0) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
     Py_RETURN_NONE;
@@ -3937,7 +4076,7 @@ PyFITSObject_read_columns_as_rec_byoffset(struct PyFITSObject* self, PyObject* a
 recread_columns_byoffset_cleanup:
 
     if (status != 0) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
     Py_RETURN_NONE;
@@ -4073,7 +4212,7 @@ PyFITSObject_read_rows_as_rec(struct PyFITSObject* self, PyObject* args) {
 recread_byrow_cleanup:
 
     if (status != 0) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
     Py_RETURN_NONE;
@@ -4144,7 +4283,7 @@ PyFITSObject_read_as_rec(struct PyFITSObject* self, PyObject* args) {
 recread_asrec_cleanup:
 
     if (status != 0) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
     Py_RETURN_NONE;
@@ -4192,7 +4331,7 @@ PyFITSObject_read_image(struct PyFITSObject* self, PyObject* args) {
 
     if (fits_get_img_paramll(self->fits, maxdim, &datatype, &naxis,
                              naxes, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -4220,7 +4359,7 @@ PyFITSObject_read_image(struct PyFITSObject* self, PyObject* args) {
     }
     if (fits_read_pixll(self->fits, fits_read_dtype, firstpixels, size,
                         0, data, &anynul, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -4379,7 +4518,7 @@ PyFITSObject_read_image_slice(struct PyFITSObject* self, PyObject* args) {
 
     if (fits_read_subset(self->fits, fits_read_dtype, fpix, lpix, step,
                          0, data, &anynul, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         goto read_image_slice_cleanup;
     }
 
@@ -4455,12 +4594,12 @@ PyFITSObject_read_header(struct PyFITSObject* self, PyObject* args) {
         return NULL;
     }
     if (fits_movabs_hdu(self->fits, hdunum, &hdutype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
     if (fits_get_hdrspace(self->fits, &nkeys, &morekeys, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -4470,7 +4609,7 @@ PyFITSObject_read_header(struct PyFITSObject* self, PyObject* args) {
         // the full card
         if (fits_read_record(self->fits, i+1, card, &status)) {
             Py_XDECREF(list);
-            set_ioerr_string_from_status(status);
+            set_ioerr_string_from_status(status, self);
             return NULL;
         }
 
@@ -4478,7 +4617,7 @@ PyFITSObject_read_header(struct PyFITSObject* self, PyObject* args) {
         // can eval in python
         if (fits_read_keyn(self->fits, i+1, keyname, value, scomment, &status)) {
             Py_XDECREF(list);
-            set_ioerr_string_from_status(status);
+            set_ioerr_string_from_status(status, self);
             return NULL;
         }
 
@@ -4505,7 +4644,7 @@ PyFITSObject_read_header(struct PyFITSObject* self, PyObject* args) {
 
                 if (fits_read_key_longstr(self->fits, keyname, &longstr, comment, &status)) {
                     Py_XDECREF(list);
-                    set_ioerr_string_from_status(status);
+                    set_ioerr_string_from_status(status, self);
                     return NULL;
                 }
 
@@ -4665,16 +4804,16 @@ PyFITSObject_write_checksum(struct PyFITSObject* self, PyObject* args) {
     }
 
     if (fits_movabs_hdu(self->fits, hdunum, &hdutype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
     if (fits_write_chksum(self->fits, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
     if (fits_get_chksum(self->fits, &datasum, &hdusum, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -4699,12 +4838,12 @@ PyFITSObject_verify_checksum(struct PyFITSObject* self, PyObject* args) {
     }
 
     if (fits_movabs_hdu(self->fits, hdunum, &hdutype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
     if (fits_verify_chksum(self->fits, &dataok, &hduok, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -4750,7 +4889,7 @@ PyFITSObject_where(struct PyFITSObject* self, PyObject* args) {
     }
 
     if (fits_movabs_hdu(self->fits, hdunum, &hdutype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -4761,7 +4900,7 @@ PyFITSObject_where(struct PyFITSObject* self, PyObject* args) {
     }
 
     if (fits_find_rows(self->fits, expression, firstrow, nrows, &ngood, row_status, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         goto where_function_cleanup;
     }
 
@@ -4834,7 +4973,7 @@ PyFITS_get_keytype(PyObject* self, PyObject* args) {
 
 
     if (fits_get_keytype(card, dtype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, NULL);
         return NULL;
     } else {
         return Py_BuildValue("s", dtype);
@@ -4856,7 +4995,7 @@ PyFITS_get_key_meta(PyObject* self, PyObject* args) {
     keyclass=fits_get_keyclass(card);
 
     if (fits_get_keytype(card, dtype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, NULL);
         return NULL;
     }
 
@@ -4901,11 +5040,11 @@ PyFITS_parse_card(PyObject* self, PyObject* args) {
     if (keyclass != TYP_COMM_KEY && keyclass != TYP_CONT_KEY) {
 
         if (fits_get_keyname(card, name, &keylen, &status)) {
-            set_ioerr_string_from_status(status);
+            set_ioerr_string_from_status(status, NULL);
             goto bail;
         }
         if (fits_parse_value(card, value, comment, &status)) {
-            set_ioerr_string_from_status(status);
+            set_ioerr_string_from_status(status, NULL);
             goto bail;
         }
         if (fits_get_keytype(value, dtype, &status)) {
@@ -4914,7 +5053,7 @@ PyFITS_parse_card(PyObject* self, PyObject* args) {
                 is_undefined=1;
                 status=0;
             } else {
-                set_ioerr_string_from_status(status);
+                set_ioerr_string_from_status(status, NULL);
                 goto bail;
             }
         }
@@ -4965,6 +5104,7 @@ static PyMethodDef PyFITSObject_methods[] = {
 
     {"reshape_image",          (PyCFunction)PyFITSObject_reshape_image,          METH_VARARGS,  "reshape_image\n\nReshape the image."},
     {"write_image",          (PyCFunction)PyFITSObject_write_image,          METH_VARARGS,  "write_image\n\nWrite the input image to a new extension."},
+    {"write_subset",         (PyCFunction)PyFITSObject_write_subset,         METH_VARARGS,  "write_subset\n\nWrite a rectangular subset to the image."},
     //{"write_column",         (PyCFunction)PyFITSObject_write_column,         METH_VARARGS | METH_KEYWORDS, "write_column\n\nWrite a column into the specified hdu."},
     {"write_columns",        (PyCFunction)PyFITSObject_write_columns,        METH_VARARGS | METH_KEYWORDS, "write_columns\n\nWrite columns into the specified hdu."},
     {"write_var_column",     (PyCFunction)PyFITSObject_write_var_column,     METH_VARARGS | METH_KEYWORDS, "write_var_column\n\nWrite a variable length column into the specified hdu from an object array."},